feat: add protocol for network error presentation
